moved styled icons for map.js to their own files map.styles.js relates to issue #7
