+++ src/components/Map.js
@@ -4,7 +4,14 @@ import { Map, Marker, Popup, TileLayer, Tooltip } from "react-leaflet";
 import MarkerClusterGroup from "react-leaflet-markercluster";
 import Icon from "./MarkerIcon";
 import L from "leaflet";
-import styled from "styled-components";
+import {
+  PopupInfo,
+  PopupLabel,
+  CenteredSection,
+  Pill,
+  Button
+} from "./map.styles";
+// import styled from "styled-components";
 import "./Map.css";
 
 type Props = {|
@@ -32,19 +39,8 @@ type Props = {|
 
 type MarkerData = {| ...Props, key: string |};
 
-const useClassColor = {
-  A1: "#ff80cc",
-  A3: "#9eebcf",
-  B1: "#96ccff",
-  B2: "#fbf1a9",
-  B8: "#ffb700",
-  D1: "#a463f2",
-  D2: "#ff6300",
-  Other: "#fff"
-};
-
 // Date formatter //
-function formatDate(input) {
+export function formatDate(input) {
   var datePart = input.match(/\d+/g),
     year = datePart[0].substring(2), // get only two digits
     month = datePart[1],
@@ -59,31 +55,6 @@ const iconSelect = useClass =>
     className: "custom-icon",
     html: ReactDOMServer.renderToString(<Icon useClass={useClass} />)
   });
-
-const PopupLabel = styled.div.attrs({
-  className: "b mb1"
-})``;
-
-const PopupInfo = styled.div.attrs({
-  className: "mb1"
-})``;
-
-const CenteredSection = styled.div.attrs({
-  className: "center w-90 tc bt bw1 pv3 mt3 ph2"
-})``;
-
-const Pill = styled.div.attrs({
-  className: "f6 br-pill ph3 pv2 mb2 dib black b  ml-auto mr-auto"
-})`
-  background: ${props => useClassColor[props.use_class]};
-`;
-
-const Button = styled.a.attrs({
-  className:
-    "f6 grow no-underline br-pill ph3 pv2 mv2 dib link white bg-hot-pink avenir button-reset b-none"
-})`
-  color: white !important;
-`;
 
 const MarkerWithPopup = ({
   geolocation,
@@ -105,7 +76,7 @@ const MarkerWithPopup = ({
   landlord_tenants_act
 }: Props) => {
   const price = price_sqft.toFixed(2).replace(/\d(?=(\d{3})+\.)/g, "$&,");
-  // console.log(MarkerWithPopup);
+
   return (
     <Marker position={JSON.parse(geolocation)} icon={iconSelect(use_class)}>
       <Popup
